+++ ui/Main.java
@@ -1,4 +1,4 @@
-package temporaryUIPanel;
+package ui;
 
 public class Main {
     // 142 cm=426 pixels
@@ -6,6 +6,6 @@ public class Main {
     final public static int TABLE_WIDTH=336;
     final public static int TABLE_HEIGHT=672;
     public static void main(String[] args) {
-        new BilliardsFrame();
+        // new BilliardsFrame();
     }
 }
